Refactor app creation androutes into own module
* Uvicorn now runs directly on `__main__`
* Add argparser for useful cmd args such as live reload
* Move app creation to `create_app` function for reusability
* Move routes into own module. This creates a temporary (to be
  refactored) dependency injection on the `app.state` since the
  controller and template object is being attached directly. Not great
  but will change later.
